requirements.txt
psycopg2-binary >= 2.7, < 2.8
flask >= 1.0.2, < 1.1